requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.13.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
phonenumbers==8.12.41
Pillow==9.0.0
psycopg2==2.9.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
